Add subcommand & argument resolver
Given a command tree and an argv, locate the corresponding subcommand
and truncate subcommands out of the argv. That's an annoying parsing
challenge, mostly because you don't know what options accept (or
require) arguments, so you don't know if the string is an option
argument, command argument, or a subcommand.

Fun.