Add function to bring it all together
There's a new function called `resolveArgv` that does pretty much just
that. Given the root of the command tree, it takes an argv, figures out
what command it corresponds to, parses out the options, and leaves the
rest as subcommand arguments.

I think I'm just about ready to make the main entry file.