Add a ref to the parent command
I've made some small additions to the normalized command tree data
structure. Now each command has a ref to the parent command (or null),
and a field with its own name. The goal of these changes is to improve
error messages on higher-level utilities that don't readily have that
debugging information.