Add top-level config
I've been thinking about this for a while. There are other config
options I'd like to pass besides the commands, such as the command name,
package version (for help pages), a "did you mean ..." toggle, that sort
of thing.

Now the config that `createCli` accepts is nested one layer deeper, with
the top-level object reserved for more generalized configuration.