Use public option names instead of local
There's one implicit downside to `createApi(...)`: it forces your option
names to be backwards compatible. That's not a great thing because it's
not obvious unless you really think about it. Plus, it's additional
surface area to maintain.

That's why I've changed it. Now the options are named exactly as the
flags, just without the leading `--` part.