Replace hyphens in long flags with camel case
Instead of forcing people to use hyphens in the options object (e.g.
`{'long-flag': true}`), I change the long flags to camel case. There's
a small risk of namespace conflict, if the user defined `--two-words`
and `--twoWords` for the same command, but I can't think of a single
reason a human might want to do that, so I'm ignoring that edge case.