Remove support for shorthand flag "=" signs
I added a feature a while ago because I thought it would be cool. It
allowed users to pass `-p=8080` instead of `-p 8080`. In hindsight, it
broke trend with other CLI parsers. My goal is to keep dispute aligned
with common CLI patterns.

This doesn't affect `--port=8080`, that syntax is still valid and will
remain valid. Only shorthand flags are changing.