Fix the remaining TypeScript errors
The build passes! There were a few hacks here and there, but all is
well. Plus, I did a proof of concept, and TypeScript appears to be
capable of representing richer types than Flow ever could. This might
make it possible to automatically infer the type of the options object.

I still need to clean a lot of this up. But who knows when (if) that'll
happen.