Replace package.json import with require call
This is a fun one. Apparently if you `import` something in TypeScript,
the build context changes. Instead of mirroring `src/` into `dist/`,
TypeScript infers that your build context is now `dist/package.json` and
`dist/src/`. Obviously not ideal. Apparently switching that to
a `require(...)` call inline resolves the issue.