Replace usage parsing with a library
Having very few dependencies is nice. Maintaining a hand-rolled parser
is not.

I'm beginning the migration off of my hand-written usage parser onto
a parser combinator library called parjs. It's a chunkier library so
I'll need to get fancy with rollup or webpack if I want to keep a small
footprint.

I also evaluated Parsimmon and Nearley. I've used Parsimmon in the past
and I really like it, but the library is aging and has some annoyances
(e.g. it doesn't compose well and encourages the use of regular
expressions).

Nearley is more standard, but as a personal preference I can't get
behind embedding JavaScript in the nearley syntax. It looks awful,
breaks tooling, and requires a 2-phase build step. Plus I really like
the modularity of parser combinators vs the all-or-nothing BNF grammar.

That left me with parjs. Not only does it do exactly what I wish
Parsimmon could do, but it's written in TypeScript and has *phenomenal*
type inference. Those features sold me. Plus, the chunky part of parjs
is for byte parsing, a feature I don't need so theoretically I can
compile it away.