Replace parjs with parsimmon
I kept running into issues with parjs. It's 90% there, but it ain't
ready yet. The biggest problems:
- Lack of customizable parsing errors
- Poor support for manual parsing errors

To top it off, the most recent publish has been completely broken for
over a month. I like parjs better, but Parsimmon is stable, familiar,
and customizable as the day is long.

Sorry parjs.